refactor: VariableRecoveryRate to its own wrapper
